Module 1/wormhole/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.32
mysql-connector-python>=8.0.27
python-dotenv>=0.19.2